Restructure psf code in a subfolder
