+++ gammapy/irf/psf/__init__.py
@@ -1,0 +1,5 @@
+from .gauss import *
+from .kernel import *
+from .king import *
+from .psf_map import *
+from .table import *
