+++ gammapy/irf/psf/psf_map.py
@@ -4,9 +4,9 @@ import astropy.units as u
 from gammapy.maps import Map, MapCoord, WcsGeom
 from gammapy.modeling.models import PowerLawSpectralModel
 from gammapy.utils.random import InverseCDFSampler, get_random_state
-from .irf_map import IRFMap
-from .psf_kernel import PSFKernel
-from .psf_table import EnergyDependentTablePSF, TablePSF
+from .kernel import PSFKernel
+from .table import EnergyDependentTablePSF, TablePSF
+from ..irf_map import IRFMap
 
 __all__ = ["PSFMap"]
