+++ gammapy/irf/psf/tests/test_psf_gauss.py
@@ -5,7 +5,7 @@ from numpy.testing import assert_allclose, assert_almost_equal
 from astropy import units as u
 from astropy.io import fits
 from astropy.utils.data import get_pkg_data_filename
-from gammapy.irf.psf_gauss import (
+from gammapy.irf import (
     EnergyDependentMultiGaussPSF,
     HESSMultiGaussPSF,
     multi_gauss_psf_kernel,
