+++ gammapy/irf/psf/tests/test_psf_map.py
